requirements.txt
discord.py
openai
flask
gunicorn
python-dotenv